refactoring: move request handlers to api package
